added evaluate as proq method and removed execute in cli
