+++ src/proq/execute_utils.py
@@ -3,23 +3,25 @@ from concurrent.futures import ThreadPoolExecutor
 from itertools import repeat
 
 
-class BuildFailedError(Exception):
-    """Raised when a build process fails.
+class CommandFailedError(Exception):
+    """Raised when a command process fails.
 
     Attributes:
-        build_output (str): Contains the output from the build process.
+        command_output (str): contains the output from the command
     """
 
-    def __init__(self, build_output, *args):
-        super().__init__(build_output, *args)
-        self.build_output = build_output
+    def __init__(self, command_output, *args):
+        super().__init__(command_output, *args)
+        self.command_output = command_output
 
 
-def build(build_command) -> str:
-    """Builds with the given build command.
+def get_command_output(command: str, stdin: str = "", raise_on_fail: bool = False):
+    """Runs the given command and returns the output.
 
     Args:
-        build_command (str):  build command to run in a subprocess
+        command (str):  build  to run in a subprocess
+        stdin (str): the contents of the stdin passed
+        raise_on_fail (bool): whether to raise an exception on non zero return status.
 
     Return:
         output (str):  The output of build command
@@ -28,30 +30,21 @@ def build(build_command) -> str:
         BuildFailedError: if build process returns a non-zero
     """
     result = subprocess.run(
-        build_command.split(), stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True
-    )
-    build_output = result.stderr + result.stdout
-    if result.returncode != 0:
-        raise BuildFailedError(build_output=build_output)
-    return build_output
-
-
-def get_output(command: str, stdin: str):
-    result = subprocess.run(
         command.split(),
         input=stdin,
         stdout=subprocess.PIPE,
-        stderr=subprocess.STDOUT,
+        stderr=subprocess.PIPE,
         text=True,
     )
-    return result.stdout
+    output = result.stderr + result.stdout
+    if raise_on_fail and result.returncode != 0:
+        raise CommandFailedError(command_output=output)
+    return output
 
 
-def get_outputs(command, stdins: list[str]):
+def get_outputs(command, stdins: list[str], raise_on_fail: bool = False):
     n = len(stdins)
     with ThreadPoolExecutor(max_workers=n) as executor:
         return executor.map(
-            get_output,
-            repeat(command, n),
-            stdins,
+            get_command_output, repeat(command, n), stdins, repeat(raise_on_fail, n)
         )
